+++ zmq/tests/test_asyncio.py
@@ -1,6 +1,6 @@
 """Test asyncio support"""
 
 try:
-    from ._test_asyncio import TestAsyncIOSocket
+    from ._test_asyncio import TestAsyncIOSocket, TestAsyncioAuthentication
 except SyntaxError:
     pass
